news page with fake data

+++ lib/shared/models/news_model.dart
@@ -1,6 +1,6 @@
 class News {
   String id;
-  double newsId;
+  int newsId;
   String author;
   String title;
   String content;
